requirements.txt
django==4.2.29
djangorestframework==3.16.1
drf-spectacular==0.29.0
requests==2.32.5
